:sparkles: add GithubIssue object and GithubIssueRepository in domain and implementation

+++ marketplace-core/src/dto/contributions.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, JsonSchema)]
 pub struct ContributionCreation {
-	github_issue_number: u128,
-	project_id: u128,
+	github_issue_number: i64,
+	project_id: i64,
 	gate: u8,
 }
 
 impl ContributionCreation {
-	pub fn new(github_issue_number: u128, project_id: u128, gate: u8) -> Self {
+	pub fn new(github_issue_number: i64, project_id: i64, gate: u8) -> Self {
 		Self {
 			github_issue_number,
 			project_id,
@@ -17,11 +17,11 @@ impl ContributionCreation {
 		}
 	}
 
-	pub fn github_issue_number(&self) -> u128 {
+	pub fn github_issue_number(&self) -> i64 {
 		self.github_issue_number
 	}
 
-	pub fn project_id(&self) -> u128 {
+	pub fn project_id(&self) -> i64 {
 		self.project_id
 	}
